Move auth classes
